--- src/main/java/com/jjss/civideo/domain/auth/entity/BloodType.java
@@ -1,5 +1,0 @@
-package com.jjss.civideo.domain.auth.entity;
-
-public enum BloodType {
-    A, B, O, AB
-}
