--- src/main/java/com/jjss/civideo/domain/auth/entity/Gender.java
@@ -1,5 +1,0 @@
-package com.jjss.civideo.domain.auth.entity;
-
-public enum Gender {
-    MAN, WOMAN
-}
