+++ src/main/java/com/jjss/civideo/domain/user/entity/BloodType.java
@@ -1,0 +1,5 @@
+package com.jjss.civideo.domain.user.entity;
+
+public enum BloodType {
+    A, B, O, AB
+}
