+++ src/main/java/com/jjss/civideo/domain/user/entity/Gender.java
@@ -1,0 +1,5 @@
+package com.jjss.civideo.domain.user.entity;
+
+public enum Gender {
+    MAN, WOMAN
+}
