+++ src/main/java/com/jjss/civideo/domain/user/service/OAuth2AuthorizationRequestRepository.java
@@ -1,4 +1,4 @@
-package com.jjss.civideo.domain.auth.service;
+package com.jjss.civideo.domain.user.service;
 
 import com.jjss.civideo.global.util.CookieUtil;
 import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
